requirements.txt
pyTelegramBotAPI==4.23.0
requests
python-dotenv
